docs: Swagger API docs Controller로 이동
